backend/requirement.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
firebase-admin==6.1.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-cloud-vision==3.3.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
msgpack==1.0.4
numpy==1.24.2
oauth2client==4.1.3
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pymongo==3.12.3
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
shapely==2.0.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
